// src/interfaces/trailers.ts
// Generated by https://quicktype.io

export interface Trailers {
  id:      number;
  results: TrailerResult[];
}

export interface TrailerResult {
  iso_639_1:    string;
  iso_3166_1:   string;
  name:         string;
  key:          string;
  site:         string;
  size:         number;
  type:         string;
  official:     boolean;
  published_at: string;
  id:           string;
}